BIG refactor to close the client server loop
This commit makes large changes to the code base to make it so the
clients will register for specific events with the main server, and then
kick off scenes when they are triggered. The scene information was also
moved into a new file, scene-config.json.

Finally, removed some additional, but old, trigger code that wasn't used
at all