Update spooky bulb player w/ bug fix + comments
The spooky bulb player need to keep track of all ongoing bulbs to ensure
it can cancel any lights when a new event comes in for an ongoing light.
However, the map we were using stopped working, or never worked? It
seems that we were using a map interface, but there was no underlying
implementation underneath. That is no bueno. This updates the map to an
actual object and verified we are keeping track of bulbs correctly